feat: add user profile management component with form validation and API integration
- Implemented profile.component.html for user profile display and editing.
- Created profile.component.scss for styling the profile page.
- Developed profile.component.ts to handle profile loading, saving, and validation.
- Added global design tokens in _public-tokens.scss for consistent styling.
- Included Postman API collection for testing various endpoints related to user authentication, plants, cart, and orders.

+++ DigiXanh.API.Tests/Patterns/OrderProcessingFacadeTests.cs
@@ -286,7 +286,7 @@ public class OrderProcessingFacadeTests : IDisposable
             RecipientName = recipientName,
             Phone = phone,
             ShippingAddress = address,
-            PaymentMethod = PaymentMethodDto.Cash
+            PaymentMethod = (int)PaymentMethodDto.Cash
         };
 
         // Act
@@ -454,7 +454,7 @@ public class OrderProcessingFacadeTests : IDisposable
             RecipientName = "Nguyễn Văn A",
             Phone = "0900000000",
             ShippingAddress = "123 Lê Lợi, Q.1, TP.HCM",
-            PaymentMethod = method,
+            PaymentMethod = (int)method,
             ReturnUrl = "https://example.com/payment-return"
         };
     }
